+++ src/hittable.rs
@@ -6,7 +6,7 @@ use crate::vec3;
 pub struct HitRecord {
     pub p: vec3::Point3,
     pub normal: vec3::Vec3,
-    pub t: f32,
+    pub t: f64,
     pub front_face: bool,
 }
 
@@ -24,7 +24,7 @@ impl HitRecord {
 
 /// Trait which describes if a hit with a ray and returns a HitRecord
 pub trait Hittable {
-    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
+    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
 }
 
 use std::rc::Rc;
@@ -45,7 +45,7 @@ impl HittableList {
 }
 
 impl Hittable for HittableList {
-    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         let mut temp_rec = HitRecord::default();
         let mut hit_anything = false;
         let mut closest_so_far = t_max;
